refactor: Move LoginNoticeModule to modules dir
