+++ Source/Plugins/Core/com.equella.core/js/tsrc/modules/LoginNoticeModule.ts
